Updated hcomments to use django comments external app
