+++ hcomments/forms.py
@@ -1,6 +1,6 @@
 # -*- coding: UTF-8 -*-
 from django import forms
-from django.contrib.comments.forms import CommentForm
+from django_comments.forms import CommentForm
 
 from captcha.fields import ReCaptchaField
