+++ pycon/urls.py
@@ -17,7 +17,7 @@ urlpatterns = patterns('',
     (r'^admin/filebrowser/', include(fsite.urls)),
     (r'^admin/rosetta/', include('rosetta.urls')),
     (r'^admin/', include(admin.site.urls)),
-    (r'^comments/', include('django.contrib.comments.urls')),
+    (r'^comments/', include('django_comments.urls')),
     (r'^p3/', include('p3.urls')),
     (r'^conference/', include('conference.urls')),
     url(r'^conference/talks/(?P<slug>[\w-]+)$', 'conference.views.talk',
