+++ requirements.txt
@@ -57,3 +57,4 @@ stripe==1.19.1
 suds==0.4
 Unidecode==0.04.17
 django-recaptcha==1.0.5
+django-contrib-comments==1.7.1
